feat(geometry): add per-product bbox + triangle count to worst-host diagnostics
Threads the host bounding box and final triangle count already captured by
record_host_cut_effect through into the WorstHost contract (bbox/triangleCount,
camelCase), so CSG failure diagnostics carry enough per-product detail to
locate and size the offending mesh without a separate pass. Split the inline
test module out of router/diagnostics.rs into router/diagnostics/diagnostics_contract_tests.rs
to keep the file under its module-size ratchet budget (now 586 lines, down
from 710).
